feat: Refactor user profile management and enhance metrics handling
- Updated user registration to include first and last name.
- Implemented user profile management with sections for personal info, password change, and account deletion.
- Enhanced metrics handling by switching to GORM for database operations.
- Improved timeline and correlation data retrieval with optimized SQL queries.
- Added custom GORM logger for better SQL query logging.
- Updated frontend templates for user profile and registration forms to support new fields and HTMX interactions.
- Refactored various repository methods to utilize GORM's features for cleaner code.

+++ server/internal/router/middleware.go
@@ -20,7 +20,7 @@ func UserLoaderMiddleware(log *zap.Logger) gin.HandlerFunc {
 			return
 		}
 
-		user, err := repository.GetUserByID(c.Request.Context(), userID)
+		user, err := repository.GetUserByID(c.Request.Context(), uint(userID))
 		if err != nil {
 			// User ID from session is invalid (user was deleted, etc.)
 			log.Warn("Invalid user ID in session, clearing.",
@@ -36,7 +36,7 @@ func UserLoaderMiddleware(log *zap.Logger) gin.HandlerFunc {
 		}
 
 		// User is valid, store user object in context for other handlers.
-		log.Debug("User loaded from session and added to context.", zap.Int("userID", user.ID))
+		log.Debug("User loaded from session and added to context.", zap.Int("userID", int(user.ID)))
 		c.Set("user", user)
 		c.Next()
 	}
